Move full text entities from Metadata project to fuzzyjoin project go avoid Maven conflict
