+++ asterixdb/asterix-fuzzyjoin/src/main/java/org/apache/asterix/fuzzyjoin/fulltextentity/AbstractFullTextFilter.java
@@ -17,12 +17,12 @@
  * under the License.
  */
 
-package org.apache.asterix.metadata.entities.fulltext;
+package org.apache.asterix.fuzzyjoin.fulltextentity;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import org.apache.asterix.metadata.api.IFullTextFilter;
+import org.apache.asterix.fuzzyjoin.fulltextentity.IFullTextFilter;
 
 public abstract class AbstractFullTextFilter implements IFullTextFilter {
     protected final String name;
